refactor(platform): split proxy routing billing tests
